Support multiple APNs per packet (see details below)
Versions of APND prior to 0.1.4 had a bug in APND::Daemon::Protocol which
allowed only one APN to be sent per packet.

Ex:

    APND::Notification.open_upstream_socket do |sock|
      sock.write(notification1.to_bytes)
      sock.write(notification2.to_bytes)
    end

This code would cause both notifications to be sent to the APND::Daemon in
a single packet (separated by new lines \n). The first notification would be
parsed from the packet, and then the others were discarded.

This commit fixes the issue by searching each line in a received packet for
APNs to queue.